TP-60 Reorganizes component directory structure.
